Move Android AI input UI
